Clean up choosing board language
Only track when the user has overridden the settings for a board language.
This is important because Tanglet needs to change the last vowel of found
words when the board language is set to Hebrew. Players can now set a
custom word list for Hebrew and it will work correctly.

Also only track when the player has set the language to something other
than the default board language. This allows them to change locales and
have the game use the new locale settings.